requirements.txt
channels
channels_redis
celery
cryptography
Django
django-redis
dj_database_url
django-cloudinary-storage
django-cors-headers
django-csp
django-extensions
django-filter
django-ratelimit
djangorestframework
djangorestframework-simplejwt
django-allauth
dj-rest-auth
Pillow
python-decouple
openai
redis
psycopg2-binary
langdetect
stripe
isort
PyJWT
locust
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
